requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
alembic
sqlalchemy